#11 fixing segmentation error
